Add a makefile rule to setup the cluster.
The setup is supposed to happen in a declarative way, using customize.
The cluster-setup/base folder needs to contain a per-feature folder with the base manifest.
The cluster-setup/ci-cluster needs to contain the related kustomized version of the base folder.
Inside the ci-cluster folder, there must be an optional wait_for_it.sh file that must return only if the given
feature is deployed and the cluster is ready.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>