Use leader election mechanism from controller-runtime library
There are two options for the leader election. The SDK defaults to the simpler one ("Leader for life") which trades off slower failover (5 minutes by default - the pod-eviction-timeout) for a guarantee that there can never be two leaders. The other one (used here) fails over faster, but there may be two leaders if one node is running a clock 50% faster(!) than the others unless you tweak the default timeouts to accommodate the clock skew rate you expect.

The controller-runtime library provides a way to set leader among pods via config map annotation, that includes leader lease. In case when the pod fails to update the lease under the config map, the other pod will acquire the leader.

Otherwise, the operator-SDK do not provide the lease feature and just creates the
config map, so until the config map exists, another pod can not acquire the leader,
but the config map will exist as long as pod exists(also in terminating state),
so it will prevent HA in cases when node not responsive for some reason(network, kernel panic, ...).

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>