Ensure 'replaces' field is empty if the CSV does not replace a previous version
The 'replaces' field must be set to a valid entry in the registry container
if it is set. Otherwise if 'replaces' does not match a CSV in the container
the registry database building will fail at container build time. Replaces
has to be empty if there is no previous CSV version being replaced.

Signed-off-by: David Vossel <davidvossel@gmail.com>