Don't use hardcoded CI registry names
This unfortunately means we can't use the registry image in CI, because we don't have
a good env var available in the image build step, so we can't set the operator image
in the registry image. Alternatively we deploy the CRDs and CSV directly now in CI.

For manual tests with a registry image you can override FEATURES_ENVIRONMENT to
"other-clusters" when deploying.

Also, the upstream registry image Dockerfile was prepared to tag images based on
branch names, like branch release-4.4 results in tag v4.4.

Signed-off-by: Marc Sluiter <msluiter@redhat.com>