tools: add a simpler tool to update the image ref
This patch adds a tool to just and only update the image references
in a CSVs.
We want to use this tool to support the upgrade flows when
we will have more than one stable release.
This way, we will be able to update the older CSVs automatically
from a newer release with minimal changes.

Say, the manifest image of version N could update the bundled
manifests of version N-1 with minimal changes, thus simplifying
the upgrade graphs.

The intended overall flow is the following:
When we release version Y, we will include manifest bundles for older
versions (version Y-1 to say the least).
If a bugfix comes in stream Y-1, with this tool will have the *option*
to pull the newer version in the current Y version and release in the
manifest bundles.

Without this patch, we have only two options:
1. NEVER update the Y-1 CSVs once version Y is out. This way, should
   bugfix versions Y-1 be released, we will have no option of pulling
   them from version Y releases. This is a fine choice per se
    and up to us to decide if we want to go this way
2. When composing the Y version metadata images, pull the most recent
   CSVs from most recent metadata image from version Y-1.

This PR enables us to simplify option #2 while keeping option #1 totally
open.
The rationale is that the most frequent change to CSV is update of the
image reference, a change we can anticipate and it is safe to reproduce.
Everything else in the CSV changes very rarely in a stable stream.

Signed-off-by: Francesco Romani <fromani@redhat.com>