Updated OLM manifest generation
The old method with the temporary CSV in the /deploy/olm-catalog directory, and using the same directory for the registry image build, was leading to having the temporary 0.0.1 CSV in the image also when building versions != 0.0.1.
Now that directory is completely recreated each time with operator-sdk, and the final manifests for the registry image are copied to / enriched in /build/_output/olm-catalog. The enriched dev CSV is copied back to /deploy/olm-catalog (for tracking it in git).
Additionally the package channel now matches the CSV version, in preparation for having 2 versions in it for testing updates.

Signed-off-by: Marc Sluiter <msluiter@redhat.com>