must-gather: collect basic HW node topology
This patch enhance must-gather to collect basic information about
the HW topology of the nodes. We use standard system utilities like
`lspci -nvv` and `lscpu -e`.
We use the same approach pioneered in cnv-must-gather to collect
data from worker nodes.

When we reach the `gather_node` stage, we create a daemonset which
will run on each worker nodes. Thanks to this daemonset, we create
hollow empty privileged pods. Privileged pods are needed to be able
to access the worker node `/proc` and `/sys` pseudofs.

Once the daemonset is ready, we iterate over all the worker nodes
and we run our collection programs (lscpu, lspci...) in the privileged
pods.
When we collected all the information we need, we clean up after
ourselves removing the daemonset with all the dependencies.

This approach is complex but minimizes the time on which privileged
pods run in the cluster.

Signed-off-by: Francesco Romani <fromani@redhat.com>