Deploy the crio config snippet and a new runtime class via the operator
Functionality to disable/enable the CPU load balancing will be implemented on the CRI-O level,
the code under the CRI-O will disable/enable CPU load balancing only when:
- the pod uses `performance-<profile_name>` runtime class
- the pod has `cpu-load-balancing.crio.io: true` annotation

The performance-addon-operator will be responsible for the creation of the high-performance runtime handler config snippet,
it will have the same content as default runtime handler, under relevant nodes,
and for creation of the high-performance runtime class under the cluster.

A user will be responsible for specifying the relevant runtime class and annotation under the pod.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>