assets/tuned: add support for isolcpus managed_irq flag
The affinity of managed interrupts is completely handled in the kernel and
cannot be changed from user space. It can happen that a managed interrupt whose
affinity mask contains both isolated and housekeeping CPUs is routed to an isolated
CPU. As a consequence IO submitted on a housekeeping CPU causes interrupts
on the isolated CPU.

Tuned added support for the managed_irq flag of isolcpus kernel parameter that tries
to avoid the abover scenario.

The managed_irq flag works only in combination with a X-Y cpu range of isolated cpus.

Use the new flag to avoid latencey spikes caused by interrupts on isolated CPUs.

Signed-off-by: Marcel Apfelbaum <marcel@redhat.com>