metadata: add metadata image, to be used with OPM
Not long ago the OLM gained support for the OPM tooling, which changes
the way we pack and deliver the metadata to be consumed by the OLM
itself. The new, recommended flow, makes use of the so-called "operator
bundles" and of the "index images".

See: https://github.com/operator-framework/operator-registry/blob/master/docs/design/operator-bundle.md

For our intents and purposes, the main benefit here is that we don't
need to maintain anymore binary registry images, but rather only
much simpler and less challenging metadata images.

Signed-off-by: Francesco Romani <fromani@redhat.com>