Makefile: metadata: fix pullspec out of container
When we build the registry-container, we fix the CSV pullspec with the
correct URL of the freshly-built container image inside the Dockerfile.
This works because the registry image is based on ose-operator-registry,
which have the core unix tools inside.

When we create the metadata container we can't do the same, because
the metadata image is FROM scratch, and it must be kept that way.
Multi-stage builds are also unsupported -and they will unlikely be
supported ever.

Thus the way forward is to replace the pullspec placeholder with the
correct pullspec before to even start the container creation.

Signed-off-by: Francesco Romani <fromani@redhat.com>