scripts: add target to generate the manifests
Up until now, we fixed the manifests pullspec inside the relevant
Dockerfile, most prominent example is the registry dockerfile.
For the metadata image, we can't do that: the Dockerfile starts FROM
scratch.

Doing it in the relevant makefile target turns out to be a great idea
either, because we want it to be a separate -and composable- step.
IOW fixing manifests should not be tied to (re)generate them from dev.

So this PR moves the manifest generation (which ATM just means "fixing
the pullspec") in its separate composable target.
In order to avoid any disruption to the existing flows, we generate the
manifests intentionally in another subtree under `build/_output`, and we
consume it in the metadata container; the existing targets aren't
affected at all.

Since we planm to phase out the registry images soon enough, we will
later just clean up the registry targets and keep this new flow only.

Signed-off-by: Francesco Romani <fromani@redhat.com>