bundle: fix OLM manifest tree
With this patch we make sure the manifests tree we create is
compliant with OLM layout.

For the transitional period, we fix the layout only for OLM-compliant
versions, which in turn we detect checking the existence of the OLM
metadata. Any other version is taken verbatim from the master manifest
subsource tree (deploy/olm-catalog).

In the future, when we will be able to drop the appregistry support,
we will cleanup and rearrange the master manifest subsource tree,
simplifying again the `generate-manifests-tree.sh` script.

Signed-off-by: Francesco Romani <fromani@redhat.com>