index: fix building of the index images
It seems that at least from version "18.02.0-ce"
the `initializer` found in the
`quay.io/operator-framework/upstream-registry-builder` image requires a
more precise path specification to read the manifests.

We fix our scripts to comply.
This isse was not triggered earlier on CI because we don't build *index*
images in CI, in turn caused by current CI limitations; we only build
bundle images in CI.

Signed-off-by: Francesco Romani <fromani@redhat.com>