operator-sdk: bump to v0.18.2
We update to v0.18.2 to make sure we have `apiextensions.k8s.io/v1` for
CRDs. We cannot jump yet to 1.0.0 because v1.0.0 introduces too many
potentially breaking changes which er ned to evaluate.

Moving from 0.15.z to 0.18.z is a big step forward in modernization
anyway. The next jump will be to 1.0.0.

Steps:
1. https://sdk.operatorframework.io/docs/upgrading-sdk-version/version-upgrade-guide/#v016x
2. https://sdk.operatorframework.io/docs/upgrading-sdk-version/v0.18.0/

As part as this PR, we begin a new layout for the manifests.
Manifests under `deploy/` should change rarely, and explicitely.
All the manipulations we do to manifests should be done under
`build/_output`.

Signed-off-by: Francesco Romani <fromani@redhat.com>