hack: Update the csv-generate hack
Use kustomize and the new "generate packagemanifests"
arguments of the operator-sdk which are pretty different
fromt the prev version.

This represents the biggest logical change when moving to
operator-sdk v1.0.0.

The CSV is generated from scratch and not based on a previous copy:
    $KUSTOMIZE build config/default | $OPERATOR_SDK generate packagemanifests \
    --version="${CSV_VERSION}" \
    --channel="${CSV_CHANNEL}" \
    --default-channel=true \
    --update-objects \
    --from-version="${CSV_FROM_VERSION}" \
    --deploy-dir="deploy/olm-catalog"   \
    --output-dir="${OUT_DIR}" \
    --crds-dir="config/crd/bases"

The CRD is also “generated”, the “base” is expected to be present
The CRD is not generated twice as before, is saved into a “temp dir”,
then “recovered” since the csv-generator deletes it.

The deploy/olm-catalog/[...]package.yaml is still edited manually

Signed-off-by: Marcel Apfelbaum <marcel@redhat.com>