Fix test: "Verify rcu_nocbs kernel argument on the node"
A given process may end up being assigned in a cpu not belonging to the
reserved cpu set (i.e. if the host has more than reserved + isolated cpus). This
leads to flakes. Here we just check that cpu does not belong to the isolated
cpus because there is no guarantee for belonging to the reserved cpus.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>