Add a loop to the hugepages allocation script
The allocation of hugepages dynamically via the allocation
the script sometimes failed(the reason still under the investigation)
and does not allocate the specified number of hugepages.

And on the second try, the allocation succeeds. To try allocation
again we added the while loop that will try to allocate again for
the whole minute with the frequency one sample per second.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>