e2e: fix the runtime class test
Under our CI it happens pretty often that the test
does not get the path to the container cpuset cgroup,
probably it takes some time to sync cgroup changes between host
and machine-config-daemon container.

To see if it indeed the case I will add `Eventually` that will try
to fetch container cpuset cgroup for two minutes with 30 seconds samples.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>