Don't check for errors while waiting for profile to be applied.
In order to make the tests work with a single node cluster, every time
there is a chance the node is rebooted (after applying / deleting a profile), we just need to
wait for conditions without checking the errors. This is because if polling happens when
the cluster is down, the api calls will fail.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>