must-gather: wait when gathering node logs
In must-gather, we use a shell for loop to gather
the node logs for all the relevant units.
We run each gather process in background, which is a good
thing to speed up the task.

However, we need to wait for these processes to complete
before we move forward, otherwise we will terminate discarding
the background process data.

Bug-Url: https://bugzilla.redhat.com/1906355
Signed-off-by: Francesco Romani <fromani@redhat.com>