machineconfig: fix the RPS boot failures
It turns out the veth devices are renamed as soon as they are created,
resulting in a race between that time and the time the update-rps
service handles the interface.

If the net device directory does not exist assume the net device
has already been renamed and trace it using the systemd device
dependencies.

Signed-off-by: Marcel Apfelbaum <marcel@redhat.com>