Revert the change of the generated runtime class name
It was not a good idea to change the name of the generate runtime class
because the name is immutable field and during the upgrade you will have
failure:

```
2020-12-17T07:07:31.743Z	ERROR	controller	Reconciler error	{"reconcilerGroup": "performance.openshift.io", "reconcilerKind": "PerformanceProfile", "controller": "performanceprofile", "name": "ci-upgrade-test", "namespace": "", "error": "RuntimeClass.node.k8s.io \"performance-ci-upgrade-test\" is invalid: handler: Invalid value: \"low-latency\": field is immutable"}
```

To make it possible to change the name, we should first delete the runtime class and after the
performance-addon-operator will re-create it with the updated name.

I will create a separate PR for it, but to fix the upgrade-test I will revert the runtime-class name.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>