upgrade test: fix initial operatorgroup
when upgrading from 4.6 to 4.7, we need to make sure
to deploy the STARTING (not the ENDING) operator group,
because the operator group spec itself is one of the things
which changed during this transition, the 4.7 being more
restrictive (OwnNamespace and SingleNamespace installModes no longer
allowed).

Please note this is relevant ONLY for this transition; all the other
transitions are perfectly fine using either OperatorGroup definition.

IOW:
- 4.4 -> 4.5
- 4.5 -> 4.6
- ...
- 4.7 -> 4.8 (expected)

all use the *same* definition for both source and target OperatorGroup,
so using either to deploy first is fine.

Please note all of this is relevant only for upgrade tests.

Signed-off-by: Francesco Romani <fromani@redhat.com>