gather-sysinfo: support collecting from chroot
add `-root` option to be able to collect data from chroot-mounted
sysfs/procfs, as long as the chroot is common, which is usually
the case inside the containers.

As side effect, the list of collected files changes, but we
files we omit now are not relevant for the KNI use case.

Signed-off-by: Francesco Romani <fromani@redhat.com>