Generate a machine config name with the numerical prefix
Currently, the operator generates a MachineConfig with the name of performance-<pao-name> and
does not follows a convention of pretending a numerical value to the name so when the rendered
version of the MCP is generated it is applied in whatever alphanumerical order it shows in the
list of oc get mc.
Now, when a cluster-admin wants to do further customizations, it cannot longer create MC with
a prepend index like 99-my-custom-mc-fix as it will not be applied correctly when the MCP is rendered.

Fixes #490

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>