Update discovery mode performance profile selector to choose by NODES_SELECTOR environment variable
Currently when running cnf-tests in discovery mode the performance
profile selector is choosing the profile in the cluster with the
most nodes using it.
When running in a cluster with more than one performance profile,
the profile selector should choose the performance profile by
NODES_SELECTOR env var if it is set.