must-gather: sysinfo: make sure we copy ghw files
ghw creates snapshots in two steps:
1. build the file list of all the files to collect
2. actually create the snapshot using the list above

The step #2 cannot handle duplicates: if a I/O operation
files during the copy, the whole operation aborts.
This behaviour will maybe changed in the next versions,
but for the time being, we need to make sure to not feed
ghw with duplicate files.

The content KNI wants has a pretty large but not complete overlap
with the content ghw wants. The list of KNI pseudofiles, thus,
do include some possible overlaps.
Up until now, we solved this issue avoiding to call the functions
which created the ghw-specific filelist; we had a filelist which
included all the ghw content, plus KNI's.

This approach however proven too hard to maintain. So with this
patch we build on ghw filelist and we add the KNI content, taking
care of deduplicating entries. This will make sure no duplicated
content ever reaches ghw.

Signed-off-by: Francesco Romani <fromani@redhat.com>