Load GHW snapshots for targetted nodes
This patch loads the GHW snapshots of the targetted nodes.
This information is obtained from the must gather directory.
This patch shows examples of how the loaded
snapshots can be used to show node's CPU and Topology
Info.

In addition to the above, we have unit tests to validate
the behaviour of this patch.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>