must-gather: rename variable for clarity
must-gather already depends on the perf-node-gather ds to detect
nodes on which it should collect data.
In turn, the daemonset doesn't explicitely select any node: we run, by
design, on each schedulable node, which, by default, exclude
control-plane nodes using NoSchedule taints.

On single-node deployment, the only prerequisite is that the node
must NOT have such a taint - which is a precondition to run any useful
workload anyway.

This PR hence just renames an internal variable for clarity; no changes
in logic are actually needed.

Signed-off-by: Francesco Romani <fromani@redhat.com>