E2E test for consume NTO's per-node Tuned profile application status
The test will create a TuneD object with a TuneD profile that doesn't exist.
Tuned daemon will fail to apply it, therefore it will get into a Degraded state which eventually should be reflected in PAO status.
In the end, we remove the bad Tuned object and wait for PAO status condition to be Available again.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>