profile-creator: Don't add the 'userLevelNetworking' field to the performance profile by default
Do not automatically add the net.userLevelNetworking field to the performance profile output
if the user does not specify the corresponding user-level-networking param.

The benefit is to have a more concise profile and the user being able to
understand what fields were 'touched'.

Update the regression tests to check for both scenarios (param specified or not).

Signed-off-by: Marcel Apfelbaum <marcel@redhat.com>