e2e: cnt-hk-cpus: add test entrypoint
We want the "housekeeping processes" inside a container,
so the entrypoint and anything we run with `oc exec` to
always execute in a well defined set of container cpus,
which we call "container HK cpus" (lacking better name).

This is not the cleanest behaviour, but it represents
an established usage pattern.

Signed-off-by: Francesco Romani <fromani@redhat.com>