e2e: cluster: add cpu availability check
The cpu placement logic may require a minimum number of cores to have
a meaningful test, but constrained CI envs may not have worker with
so many usable cores for exclusive allocations.

Additionally, we have a global toggle (environment variable) to
skip or fail the test should the preconditions not be met.
- on CI, to have false negatives or permafails, we should have the flag
  disabled, hence the test should skip.
- on real envs (or on bugger CIs :) ) we should set the flag to on,
  and let test fails if the cluster doesn't met the preconditions,
  because it is supposed to be!

Signed-off-by: Francesco Romani <fromani@redhat.com>