Set maximumLatency variable
In the last update, setting the maximumLatency via the environment variables value was skipped (bug).
This update will set the values correctly.
Moreover, this update introduces a new unified environment variable for the maximum latency that will allow the user to use the same environment variable for all tests.
The specific environment variables will have precedence over this unified one.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>