tests: images prepull using daemonsets
In order to pre-pull container images on the worker nodes, we create
bogus, extremely cheap daemonsets which pull the images as side effect
in order to run. We just sleep in these daemonsets, so they are as cheap
as a container can be.

We run these daemonsets in the testing namespace, so, once the tests are
all done, the daemonsets will be cleaned up explicitly; no explicit
cleanup (albeit nice) is required.

Other alternatives are running jobs to pull images as side effects, but
they don't seem to be much simpler than daemonsets.

Daemonsets need to be named explicitely from the caller, which must also
ensure their name is unique. This is awkward, but automatic naming would
make the code too complex for the task at hand: there is no immediate
need, because the number of images we should pre pull is (and should be
kept) limited.

Finally, it's worth pointing out that this solution, while it nicely
reuses the kubernetes primitives and follows some already established
pattern in the community isn't too great, becausthere pre pulling images
should be done as part as the cluster preparation stage; so, it is
better done in the test setup stage, not in the test running stage.

Add a new tool to run the prepulling code out of the tests.

Signed-off-by: Francesco Romani <fromani@redhat.com>