Latency e2e: Verify node has sufficient cpu
The machines on which CI runs have few cpus (1-2 typically) for cost/scalability reasons.
Our tests want to exclusively allocate 1 or more cpus, which is impossible or impractical.
Trying to run our tests in these machines will lead to false negatives.

To keep the lane reliable, we add a check for worker nodes to verify nodes have sufficient
cpu amount to run the tests. Should not be the case, we skip the test (instead of failing).