Makefile: modify suits for target functests-latency-testing
Latency tests will fail if discovery mode is disabled and performance profile is not found on the cluster.

To allow proper latency testing, we add functests/0_config suit to ginkgo suits of the target functests-latency-testing, a suit that will handle creating the profile in case it doesn't exist or continue executing the tests directly if exists.