profilecreator: fix misuse of builder.Result()
The bump to kube 1.23 exposed a latent bug in how we use the
cpuset.Builder.Result() method.

Per doc, the intent of the method was to finalize the builder
and to return the built cpuset. Any subsequent call of Add()
was expected to be a NOP.

Due to a previous bug, this behaviour was not enforced, so
calls to Add() after Result(), albeit wrong, were working.

We use like this in our code because we want to accumulate CPUs
only up a threshold. So the only option is to carry a separate
counter, which allows us to efficiently accumulate as much as
we need and to call Result() exactly once, as it should be.

Signed-off-by: Francesco Romani <fromani@redhat.com>