tuned base profile to support new scheduler nobs
tuned was updated to handle new kernels (5.13 and newer)
they moved some sched_ and numa_ knobs from
the sysctl to the debugfs, thus add and abstract these knobs under the
scheduler plugin. With help of this abstraction it will write
the tuning to the correct place according to the kernel used.

Example:
[scheduler]
sched_migration_cost_ns = 500000

Will work on the old kernel the same way as:
[sysctl]
kernel.sched_migration_cost_ns = 500000

all occurences of sched_ and numa_ knobs were added
under the [scheduler] section in the base tuned profile in PAO

Signed-off-by: Yanir Quinn <yquinn@redhat.com>