e2e: latency: skip the test when requested cpus number is odd
Following the change done in this PR https://github.com/openshift-kni/performance-addon-operators/pull/830, odd cpus number will lead the latency pod to fail with SMT Alignment Error.
Add check to skip the test if the requested number of CPUs is odd number, and add a test in the latency testing suite to verify the change.