images: change ImagePullPolicy to default
When ImagePullPolicy is defined as Always, K8s always tries pulling the image even if the image exists on the machine, this obstructs the test on disconnected environments and lead to failure.
Remove this requirement from the container spec switching to the default pull policy which is IfNotPresent, allowing the pod to use the image existing on the machine.

Signed-off-by: shereenH <shajmakh@redhat.com>