e2e: Latency: add test to run oslat with less CPUs than allowed
Add new test that runs Oslat test with CPUs number = 1. Achieve that by modifying the performance profile isolatedCpus to 1.
Note that this cannot be tested by using the current describeTable entries because passing "1" as the testCpus would result with a different skip message (skipOddCpuNumber). Passing 0 is not an option because it is invalid.

Signed-off-by: shereenH <shajmakh@redhat.com>