Improve pod information to check IRQ Load Balancing
Added "resources.requests.cpu" and "resources.limits.cpu" information
for each of the containers on each pod.
This information is required to look for exclusively assigned CPUs.

CPUs are assigned to be used exclusively only to those containers
belonging to a pod with a "Guaranteed" qosClass and with an integer cpu
request
Thats why we need to add this information to must-gather.