Modify regex to reflect new "data"
Due to some dependency update profile.spec.data field is now written as
a multiple lines string instead of as a single line string.
We are using regex to check for some content in this field and these new
line feeds could make them fail so we have updated the regex to check
only the data we are interested on despite what could be in between.