e2e: latency: replace BeforeSuite and AfterSuite with functions
Replace BeforeSuite and AfterSuite with setup() and teardown functions respectively. This enables us to skip the suite at earlier stage before the setup steps.

Note that it isn't possible to skip the whole suite (and not only specific tests) from inside BeforeSuite, and the only way to do that is via t.Skip in the TestXX(t ..).

Skipping the suite will follow in the next commit.

Signed-off-by: shereenH <shajmakh@redhat.com>