e2e: latency: skip the tests earlier in the suite
This suite works properly only if the test executable exists. Currently each test checks for the existence of this common executable.
The suite performs some setup steps that are relevant for this set of tests, without considering whether the file exists or not. To avoid extra unneeded execution time, skip the suite at an early stage if the executable file does not exist.

Signed-off-by: shereenH <shajmakh@redhat.com>