must-gather:podinfo: Add RBAC elements to list pods
podinfo command need to list all pods in all namespaces scheduled to a
certain node. In order to do that it needs to have the proper
permissions.

Add a new ClusterRole,adding read permisions over pods in all namespaces,
and a ClusterRoleBinding to bind those persmissions to the
ServiceAccount used by the must-gather daemonset pods.