Avoid error messages when NRO is not installed
must-gather is trying to get NRO related information (CRDs, and
namespaced artifacts) no matter what and so, when NRO is not installed
in the cluster some nasty error messages appear because those elements
are not found.

These are not errors but to avoid missleading the users it is better to
avoid them by not gathering the NRO information if we cannot detect the
NRO namespace.
A small message appears in console in this case.